requirements.txt
altgraph==0.16.1
beautifulsoup4==4.7.1
bottle==0.12.16
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.10.3
Eel==0.10.4
Faker==0.9.1
Flask==1.0.3
future==0.17.1
get==2019.4.13
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gTTS==2.0.3
gTTS-token==1.1.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
macholib==1.11
MarkupSafe==1.1.1
mixer==6.1.3
numpy==1.16.4
pandas==0.24.2
pefile==2019.4.18
pipenv==2018.11.26
post==2019.4.13
psycopg2==2.8.2
public==2019.4.13
pycparser==2.19
pygame==1.9.6
PyInstaller==3.4
python-dateutil==2.8.0
python-vlc==3.0.7110
pyttsx==1.1
pytz==2019.1
pywin32-ctypes==0.2.0
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
virtualenv==16.6.1
virtualenv-clone==0.5.3
Werkzeug==0.15.4
whichcraft==0.5.2
whitenoise==4.1.2
